requirements.txt
flask
python-docx
openpyxl
pillow
